requirements.txt
torch~=1.13.1+cu116
numpy~=1.22.4
tqdm~=4.66.1
Pillow~=9.3.0
torchvision~=0.14.1
opencv-python~=4.8.0.76
PyYAML~=6.0.1